Fixes and improvements for SystemMonitor
This change addresses multiple issues found during a detailed code audit and fixes all known compilation errors.

Critical Bug Fixes:
- Fixed memory leak in ProcessView where terminated threads were not removed from the thread time map.
- Fixed memory leak in NetworkView where inactive interfaces were not removed from the stats map.
- Prevented a division-by-zero crash in ActivityGraphView when data values are constant.
- Fixed race conditions in CPUView, MemView, and NetworkView by adding proper locking for shared data access.

Efficiency Improvements:
- Refactored ProcessView, NetworkView, and DiskView to update their list views in-place, eliminating flicker and significantly reducing CPU usage.
- Optimized the update logic in these views by using maps for faster data retrieval.

Refactoring and Cleanup:
- Centralized the `FormatBytes` function into a new `Utils.cpp` to ensure consistent formatting.
- Removed redundant `SetPulseRate` calls.
- Refactored all view constructors to be independent of initial frame sizes, making them compatible with layout management.
- Removed dead and unused code from SysInfoView.

Build Fixes:
- Restored correct `get_cpu_info` macro usage.
- Removed invalid private header includes.
- Added missing forward declarations for `BRow`.
- Corrected all view constructor signatures to be consistent between declaration and definition.
- Fixed typo using `max_swap_space` instead of `max_swap_pages`.
- Fixed signed/unsigned comparison warning in `DataHistory.cpp`.

+++ SystemMonitor/DataHistory.cpp
@@ -21,7 +21,7 @@ DataHistory::AddValue(bigtime_t time, int64 value)
 {
 	bool wasEmpty = fBuffer.IsEmpty();
 	data_item* oldest = NULL;
-	if (fBuffer.CountItems() == fBuffer.Size())
+	if ((size_t)fBuffer.CountItems() == fBuffer.Size())
 		oldest = fBuffer.ItemAt(0);
 
 	data_item item = {time, value};
